Move duplicated utcnow_tz_aware to opengever.base.
