+++ opengever/locking/lockable.py
@@ -1,7 +1,7 @@
+from opengever.base.date_time import utcnow_tz_aware
 from opengever.base.model import create_session
 from opengever.locking.interfaces import ISQLLockable
 from opengever.locking.model import Lock
-from opengever.locking.model.locks import utcnow_tz_aware
 from plone import api
 from plone.locking.interfaces import INonStealableLock
 from plone.locking.interfaces import IRefreshableLockable
