+++ opengever/locking/tests/test_locks.py
@@ -1,7 +1,7 @@
 from datetime import timedelta
+from opengever.base.date_time import utcnow_tz_aware
 from opengever.core.testing import MEMORY_DB_LAYER
 from opengever.locking.model import Lock
-from opengever.locking.model.locks import utcnow_tz_aware
 from plone.app.testing import TEST_USER_ID
 from sqlalchemy.exc import IntegrityError
 from unittest2 import TestCase
